[MI-2232] Fixed failing tests and moved some functions to util file
